+++ src/main/java/com/becareful/becarefulserver/domain/association/dto/MemberSimpleDto.java
@@ -1,6 +1,6 @@
 package com.becareful.becarefulserver.domain.association.dto;
 
-import com.becareful.becarefulserver.domain.nursing_institution.vo.InstitutionRank;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.vo.InstitutionRank;
 import com.becareful.becarefulserver.domain.socialworker.domain.SocialWorker;
 import com.becareful.becarefulserver.domain.socialworker.domain.vo.AssociationRank;
