+++ src/main/java/com/becareful/becarefulserver/domain/auth/handler/CustomSuccessHandler.java
@@ -5,7 +5,7 @@ import com.becareful.becarefulserver.domain.auth.dto.response.OAuth2LoginRespons
 import com.becareful.becarefulserver.domain.auth.dto.response.RegisteredUserLoginResponse;
 import com.becareful.becarefulserver.domain.caregiver.domain.Caregiver;
 import com.becareful.becarefulserver.domain.caregiver.repository.CaregiverRepository;
-import com.becareful.becarefulserver.domain.nursing_institution.vo.InstitutionRank;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.vo.InstitutionRank;
 import com.becareful.becarefulserver.domain.socialworker.domain.SocialWorker;
 import com.becareful.becarefulserver.domain.socialworker.domain.vo.AssociationRank;
 import com.becareful.becarefulserver.domain.socialworker.repository.SocialWorkerRepository;
