+++ src/main/java/com/becareful/becarefulserver/domain/caregiver/domain/Caregiver.java
@@ -5,7 +5,7 @@ import static com.becareful.becarefulserver.global.constant.StaticResourceConsta
 import com.becareful.becarefulserver.domain.caregiver.domain.vo.Address;
 import com.becareful.becarefulserver.domain.caregiver.domain.vo.CaregiverInfo;
 import com.becareful.becarefulserver.domain.common.domain.BaseEntity;
-import com.becareful.becarefulserver.domain.common.vo.Gender;
+import com.becareful.becarefulserver.domain.common.domain.Gender;
 import jakarta.persistence.*;
 import java.time.LocalDate;
 import java.util.List;
