+++ src/main/java/com/becareful/becarefulserver/domain/caregiver/domain/WorkApplicationWorkLocation.java
@@ -1,7 +1,7 @@
 package com.becareful.becarefulserver.domain.caregiver.domain;
 
 import com.becareful.becarefulserver.domain.common.domain.BaseEntity;
-import com.becareful.becarefulserver.domain.common.vo.Location;
+import com.becareful.becarefulserver.domain.common.domain.vo.Location;
 import com.becareful.becarefulserver.domain.work_location.domain.WorkLocation;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
