+++ src/main/java/com/becareful/becarefulserver/domain/caregiver/dto/response/WorkScheduleResponse.java
@@ -3,7 +3,7 @@ package com.becareful.becarefulserver.domain.caregiver.dto.response;
 import static java.time.format.DateTimeFormatter.ofPattern;
 
 import com.becareful.becarefulserver.domain.common.domain.CareType;
-import com.becareful.becarefulserver.domain.common.vo.Gender;
+import com.becareful.becarefulserver.domain.common.domain.Gender;
 import com.becareful.becarefulserver.domain.matching.domain.CompletedMatching;
 import java.util.List;
 import lombok.AccessLevel;
