+++ src/main/java/com/becareful/becarefulserver/domain/common/domain/Gender.java
@@ -1,0 +1,6 @@
+package com.becareful.becarefulserver.domain.common.domain;
+
+public enum Gender {
+    MALE,
+    FEMALE;
+}
