+++ src/main/java/com/becareful/becarefulserver/domain/common/domain/vo/Location.java
@@ -1,4 +1,4 @@
-package com.becareful.becarefulserver.domain.common.vo;
+package com.becareful.becarefulserver.domain.common.domain.vo;
 
 import jakarta.persistence.Embeddable;
 import lombok.*;
