--- src/main/java/com/becareful/becarefulserver/domain/common/vo/Gender.java
@@ -1,6 +1,0 @@
-package com.becareful.becarefulserver.domain.common.vo;
-
-public enum Gender {
-    MALE,
-    FEMALE;
-}
