+++ src/main/java/com/becareful/becarefulserver/domain/community/dto/AuthorSimpleDto.java
@@ -1,6 +1,6 @@
 package com.becareful.becarefulserver.domain.community.dto;
 
-import com.becareful.becarefulserver.domain.nursing_institution.vo.InstitutionRank;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.vo.InstitutionRank;
 import com.becareful.becarefulserver.domain.socialworker.domain.SocialWorker;
 
 public record AuthorSimpleDto(
