+++ src/main/java/com/becareful/becarefulserver/domain/matching/dto/ElderlyDto.java
@@ -1,6 +1,6 @@
 package com.becareful.becarefulserver.domain.matching.dto;
 
-import com.becareful.becarefulserver.domain.common.vo.Gender;
+import com.becareful.becarefulserver.domain.common.domain.Gender;
 import com.becareful.becarefulserver.domain.socialworker.domain.Elderly;
 import com.becareful.becarefulserver.domain.socialworker.domain.vo.CareLevel;
