+++ src/main/java/com/becareful/becarefulserver/domain/nursing_institution/domain/converter/FacilityTypeConverter.java
@@ -1,7 +1,7 @@
-package com.becareful.becarefulserver.domain.nursing_institution.converter;
+package com.becareful.becarefulserver.domain.nursing_institution.domain.converter;
 
 import com.becareful.becarefulserver.domain.caregiver.domain.converter.EnumSetConverter;
-import com.becareful.becarefulserver.domain.nursing_institution.vo.FacilityType;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.vo.FacilityType;
 import jakarta.persistence.Converter;
 
 @Converter
