+++ src/main/java/com/becareful/becarefulserver/domain/nursing_institution/dto/request/UpdateNursingInstitutionInfoRequest.java
@@ -1,6 +1,6 @@
 package com.becareful.becarefulserver.domain.nursing_institution.dto.request;
 
-import com.becareful.becarefulserver.domain.nursing_institution.vo.*;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.vo.FacilityType;
 import jakarta.validation.constraints.*;
 import java.util.*;
