+++ src/main/java/com/becareful/becarefulserver/domain/socialworker/dto/SocialWorkerSimpleDto.java
@@ -1,7 +1,7 @@
 package com.becareful.becarefulserver.domain.socialworker.dto;
 
-import com.becareful.becarefulserver.domain.common.vo.Gender;
-import com.becareful.becarefulserver.domain.nursing_institution.vo.InstitutionRank;
+import com.becareful.becarefulserver.domain.common.domain.Gender;
+import com.becareful.becarefulserver.domain.nursing_institution.domain.vo.InstitutionRank;
 import com.becareful.becarefulserver.domain.socialworker.domain.SocialWorker;
 import com.becareful.becarefulserver.domain.socialworker.domain.vo.AssociationRank;
