+++ src/main/java/com/becareful/becarefulserver/domain/socialworker/dto/request/ElderlyUpdateRequest.java
@@ -1,7 +1,7 @@
 package com.becareful.becarefulserver.domain.socialworker.dto.request;
 
 import com.becareful.becarefulserver.domain.common.domain.DetailCareType;
-import com.becareful.becarefulserver.domain.common.vo.Gender;
+import com.becareful.becarefulserver.domain.common.domain.Gender;
 import com.becareful.becarefulserver.domain.socialworker.domain.vo.CareLevel;
 import java.time.LocalDate;
 import java.util.List;
